loads of IPC updates

+++ playground/PerformanceTest.cpp
@@ -17,7 +17,7 @@ int main(int argc, char *argv[])
     DCF::MessageBuffer buffer(1000000);
 
     std::cout << "encode x" << iterations << ": " << tf::measure<std::chrono::microseconds>::execution([&]() {
-        for (int i = 0; i < iterations; i++) {
+        for (size_t i = 0; i < iterations; i++) {
             auto msg = pool.allocate_shared_ptr();
             msg->setSubject("SOME.TEST.SUBJECT");
             float32_t t = 22.0;
@@ -32,7 +32,7 @@ int main(int argc, char *argv[])
 
     std::vector<PoolType::shared_ptr_type> decoded_messages(iterations);
     std::cout << "decode x" << iterations << ": " << tf::measure<std::chrono::microseconds>::execution([&]() {
-        for (int i = 0; i < iterations + 1; i++) {
+        for (size_t i = 0; i < iterations + 1; i++) {
             auto msg = pool.allocate_shared_ptr();
 
             const DCF::ByteStorage &storage = buffer.byteStorage();
